+++ example/lib/main.dart
@@ -1,5 +1,8 @@
 import 'dart:convert';
 import 'dart:io';
+import 'dart:isolate';
+import 'dart:ui';
+import 'dart:developer' as developer;
 
 import 'package:flutter/material.dart';
 import 'dart:async';
@@ -10,7 +13,20 @@ import 'package:flutter/services.dart';
 import 'package:loplat_plengi/loplat_plengi.dart';
 import 'package:device_info/device_info.dart';
 
-void main() {
+/// The name associated with the UI isolate's [SendPort].
+const String isolateName = 'isolate';
+
+/// A port used to communicate from a background isolate to the UI isolate.
+final ReceivePort port = ReceivePort();
+
+Future<void> main() async {
+  // Register the UI isolate's SendPort to allow for communication from the
+  // background isolate.
+  IsolateNameServer.registerPortWithName(
+    port.sendPort,
+    isolateName,
+  );
+
   runApp(const MyApp());
 }
 
@@ -26,11 +42,60 @@ class _MyAppState extends State<MyApp> {
   String _plengiStatus = "NOT_INITIALIZED";
   String _loplatResults = '';
   var getLocationButEnabled = true;
+  // The background
+  static SendPort? uiSendPort;
 
   @override
   void initState() {
     super.initState();
+    LoplatPlengiPlugin.setListener(callback);
     initPlatformState();
+  }
+
+  // The callback for our alarm
+  static Future<bool> callback(String msg) async {
+    String locationInfo = getLocationInfo(msg);
+    print('result: $locationInfo');
+
+    // This will be null if we're running in the background.
+    uiSendPort ??= IsolateNameServer.lookupPortByName(isolateName);
+    uiSendPort?.send(null);
+    return true;
+  }
+
+  static String getLocationInfo(String log) {
+    if (log == null || log.isEmpty) {
+      return '';
+    }
+    Map<String,dynamic> jsonData = jsonDecode(log);
+    String inNear = '';
+    String locationInfo = '';
+    if (jsonData['place'] != null) {
+      if (jsonData['place']['accuracy'] >= jsonData['place']['threshold']) {
+        inNear = '[IN]';
+      } else {
+        inNear = '[NEAR]';
+      }
+      locationInfo += '  $inNear${jsonData['place']['name']},'
+          '${jsonData['place']['tags']}(${jsonData['place']['loplat_id']}),'
+          '${jsonData['place']['floor']}F,${jsonData['place']['accuracy']}/${jsonData['place']['threshold']}';
+    }
+    if (jsonData['complex'] != null) {
+      locationInfo += '\n  [${jsonData['complex']['id']}]${jsonData['complex']['name']}';
+    }
+    if (jsonData['area'] != null) {
+      locationInfo += '\n  [${jsonData['area']['id']}]${jsonData['area']['tag']},${jsonData['area']['name']}';
+    }
+    if (jsonData['district'] != null) {
+      locationInfo += '\n  ${jsonData['district']['lv1_name']} ${jsonData['district']['lv2_name']} ${jsonData['district']['lv3_name']}';
+    }
+    if (jsonData['location'] != null) {
+      locationInfo += '\n  ${jsonData['location']['lat']}, ${jsonData['location']['lng']}';
+    }
+    if (jsonData['ad'] != null) {
+      locationInfo += '\n  ${jsonData['ad']['campaign_id']}, ${jsonData['ad']['title']}, ${jsonData['ad']['body']}';
+    }
+    return locationInfo;
   }
 
   // Platform messages are asynchronous, so we initialize in an async method.
@@ -85,36 +150,10 @@ class _MyAppState extends State<MyApp> {
   }
 
   updateLog(String log) {
-    initPlatformState();
-    print(log);
-    Map<String,dynamic> jsonData = jsonDecode(log);
-    String inNear = '';
-    String locationInfo = '';
-    if (jsonData['place'] != null) {
-      if (jsonData['place']['accuracy'] >= jsonData['place']['threshold']) {
-        inNear = '[IN]';
-      } else {
-        inNear = '[NEAR]';
-      }
-      locationInfo += '  $inNear${jsonData['place']['name']},'
-          '${jsonData['place']['tags']}(${jsonData['place']['loplat_id']}),'
-          '${jsonData['place']['floor']}F,${jsonData['place']['accuracy']}/${jsonData['place']['threshold']}';
+    if (log == null || log.isEmpty) {
+      return;
     }
-    if (jsonData['complex'] != null) {
-      locationInfo += '\n  [${jsonData['complex']['id']}]${jsonData['complex']['name']}';
-    }
-    if (jsonData['area'] != null) {
-      locationInfo += '\n  [${jsonData['area']['id']}]${jsonData['area']['tag']},${jsonData['area']['name']}';
-    }
-    if (jsonData['district'] != null) {
-      locationInfo += '\n  ${jsonData['district']['lv1_name']} ${jsonData['district']['lv2_name']} ${jsonData['district']['lv3_name']}';
-    }
-    if (jsonData['location'] != null) {
-      locationInfo += '\n  ${jsonData['location']['lat']}, ${jsonData['location']['lng']}';
-    }
-    if (jsonData['ad'] != null) {
-      locationInfo += '\n  ${jsonData['ad']['campaign_id']}, ${jsonData['ad']['title']}, ${jsonData['ad']['body']}';
-    }
+    String locationInfo = getLocationInfo(log);
     String formattedDate = DateFormat('MM-dd HH:mm:ss').format(DateTime.now());
     setState(() {
       _loplatResults = '$formattedDate\n$locationInfo\n\n$_loplatResults';
@@ -151,9 +190,9 @@ class _MyAppState extends State<MyApp> {
                 ElevatedButton(
                   child: const Text('Get Location'),
                   onPressed: getLocationButEnabled? () {
-                    setState(() {
+                    /*setState(() {
                       getLocationButEnabled = false;
-                    });
+                    });*/
                     Timer _timer = Timer(const Duration(seconds: 8), () {
                       setState(() {
                         if (!getLocationButEnabled) {