+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## 1.0.1
+
+* Support flutter background listener
+* Update Android Plengi to v2.1.1.3
+
 ## 1.0.0
+
 * Initial release.